coins/requirements.txt
lxml==4.1.1
requests==2.18.4
redis==2.10.6
kafka-python==1.3.5
celery==4.1.0
sqlalchemy==1.1.15
psycopg2==2.7.3.2